fixed a useless cast.
git-svn-id: svn://falconpl.org/falcon/core/trunk@1170 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/item_ext.cpp
@@ -250,7 +250,7 @@ FALCON_FUNC  mth_typeId ( ::Falcon::VMachine *vm )
       type = vm->self().dereference()->type();
    else {
       if ( vm->paramCount() > 0 )
-         type = (int64) vm->param(0)->type();
+         type = vm->param(0)->type();
       else
          throw new ParamError( ErrorParam( e_inv_params ).origin( e_orig_runtime ).extra( "X" ) );
    }
